requirements.txt
streamlit==1.29.0
openai>=2.6.0
huggingface_hub==0.19.4
Pillow==10.1.0
python-dotenv==1.0.0
requests==2.31.0

